refactor: migrate all csi2 references to mipi_csi2 and fix editable install
