+++ cocotbext/mipi_csi2/exceptions.py
