+++ cocotbext/mipi_csi2/packet.py
