+++ untested_wip/comprehensive_test.py
@@ -10,29 +10,29 @@ performance characteristics, and multi-configuration testing.
 import sys
 import time
 import asyncio
-from cocotbext.csi2 import *
-from cocotbext.csi2.config import Csi2Config, PhyType, DataType, VirtualChannel
-from cocotbext.csi2.packet import Csi2ShortPacket, Csi2LongPacket
-from cocotbext.csi2.utils import calculate_ecc, validate_checksum, generate_test_pattern, pack_raw10, unpack_raw10
-from cocotbext.csi2.exceptions import *
+from cocotbext.mipi_csi2 import *
+from cocotbext.mipi_csi2.config import Csi2Config, PhyType, DataType, VirtualChannel
+from cocotbext.mipi_csi2.packet import Csi2ShortPacket, Csi2LongPacket
+from cocotbext.mipi_csi2.utils import calculate_ecc, validate_checksum, generate_test_pattern, pack_raw10, unpack_raw10
+from cocotbext.mipi_csi2.exceptions import *
 
 def test_protocol_compliance():
     """Test CSI-2 protocol compliance features"""
     print("=== PROTOCOL COMPLIANCE TESTS ===")
-    
+
     # Test 1: ECC Generation and Validation
     print("\n1. ECC Generation and Validation")
     test_cases = [
         (0x2A, 0x0000),  # RAW8, VC=0, WC=0
-        (0x6B, 0x0500),  # RAW10, VC=1, WC=1280  
+        (0x6B, 0x0500),  # RAW10, VC=1, WC=1280
         (0xA4, 0x0960),  # RGB888, VC=2, WC=2400
         (0xFE, 0xFFFF),  # Max values
     ]
-    
+
     for data_id, word_count in test_cases:
         ecc = calculate_ecc(data_id, word_count)
         print(f"  DI=0x{data_id:02x}, WC=0x{word_count:04x} -> ECC=0x{ecc:02x}")
-    
+
     # Test 2: Checksum Validation
     print("\n2. Checksum Validation")
     test_payloads = [
@@ -41,12 +41,12 @@ def test_protocol_compliance():
         bytes([0xFF] * 1000),
         generate_test_pattern(32, 32, "ramp")
     ]
-    
+
     for i, payload in enumerate(test_payloads):
         checksum = sum(payload) & 0xFFFF
         is_valid = validate_checksum(payload, checksum)
         print(f"  Payload {i+1}: {len(payload)} bytes, checksum=0x{checksum:04x}, valid={is_valid}")
-    
+
     # Test 3: Packet Header Validation
     print("\n3. Packet Header Validation")
     packets = [
@@ -55,7 +55,7 @@ def test_protocol_compliance():
         Csi2ShortPacket.frame_end(2, 1),
         Csi2ShortPacket.line_end(3, 480)
     ]
-    
+
     for packet in packets:
         header = packet.header
         ecc_valid = header.validate_ecc()
@@ -65,7 +65,7 @@ def test_protocol_compliance():
 def test_data_formats():
     """Test different CSI-2 data formats and encoding"""
     print("\n=== DATA FORMAT TESTS ===")
-    
+
     # Test 1: RAW Data Formats
     print("\n1. RAW Data Format Support")
     raw_formats = [
@@ -74,26 +74,26 @@ def test_data_formats():
         (DataType.RAW12, 1.5),
         (DataType.RAW16, 2)
     ]
-    
+
     base_pixels = 1024
     for data_type, bytes_per_pixel in raw_formats:
         expected_bytes = int(base_pixels * bytes_per_pixel)
         print(f"  {data_type.name}: {base_pixels} pixels -> {expected_bytes} bytes")
-    
+
     # Test 2: RAW10 Packing/Unpacking
     print("\n2. RAW10 Packing/Unpacking")
     original_pixels = [i * 4 for i in range(40)]  # 40 pixels, 10-bit values
     packed_data = pack_raw10(original_pixels)
     unpacked_pixels = unpack_raw10(packed_data)
-    
+
     pack_efficiency = len(packed_data) / (len(original_pixels) * 2) * 100
     data_integrity = original_pixels == unpacked_pixels
-    
+
     print(f"  Original: {len(original_pixels)} pixels")
     print(f"  Packed: {len(packed_data)} bytes ({pack_efficiency:.1f}% efficiency)")
     print(f"  Unpacked: {len(unpacked_pixels)} pixels")
     print(f"  Data integrity: {data_integrity}")
-    
+
     # Test 3: Color Format Support
     print("\n3. Color Format Support")
     color_formats = [
@@ -103,37 +103,37 @@ def test_data_formats():
         (DataType.YUV422_8BIT, 2, "YUV 4:2:2"),
         (DataType.YUV420_8BIT, 1.5, "YUV 4:2:0")
     ]
-    
+
     for data_type, bytes_per_pixel, description in color_formats:
         print(f"  {data_type.name}: {description} ({bytes_per_pixel} bytes/pixel)")
 
 def test_performance_characteristics():
     """Test performance characteristics and bandwidth calculations"""
     print("\n=== PERFORMANCE CHARACTERISTICS ===")
-    
+
     # Test 1: D-PHY Performance Scaling
     print("\n1. D-PHY Performance Scaling")
     dphy_configs = [
         (1, 1000),
-        (2, 1500), 
+        (2, 1500),
         (4, 2500),
         (8, 3000)
     ]
-    
+
     for lanes, bit_rate in dphy_configs:
         config = Csi2Config(
             phy_type=PhyType.DPHY,
             lane_count=lanes,
             bit_rate_mbps=bit_rate
         )
-        
+
         total_bandwidth = lanes * bit_rate
         bit_period = config.get_bit_period_ns()
         byte_period = config.get_byte_period_ns()
-        
+
         print(f"  {lanes}-lane @ {bit_rate} Mbps: {total_bandwidth} Mbps total, "
               f"T_bit={bit_period:.3f}ns, T_byte={byte_period:.3f}ns")
-    
+
     # Test 2: C-PHY Performance Analysis
     print("\n2. C-PHY Performance Analysis")
     cphy_configs = [
@@ -141,21 +141,21 @@ def test_performance_characteristics():
         (2, 2000),
         (3, 2500)
     ]
-    
+
     for trios, symbol_rate in cphy_configs:
         config = Csi2Config(
             phy_type=PhyType.CPHY,
             trio_count=trios,
             bit_rate_mbps=symbol_rate
         )
-        
+
         # C-PHY encodes 2.28 bits per symbol on average
         effective_bandwidth = trios * symbol_rate * 2.28
         equivalent_dphy_lanes = effective_bandwidth / symbol_rate
-        
+
         print(f"  {trios}-trio @ {symbol_rate} Msps: {effective_bandwidth:.0f} Mbps effective, "
               f"~{equivalent_dphy_lanes:.1f} D-PHY lanes equivalent")
-    
+
     # Test 3: Frame Rate Calculations
     print("\n3. Frame Rate Analysis")
     resolutions = [
@@ -164,35 +164,35 @@ def test_performance_characteristics():
         (1920, 1080, "FHD"),
         (3840, 2160, "4K")
     ]
-    
+
     config = Csi2Config(PhyType.DPHY, lane_count=4, bit_rate_mbps=2500)
     total_bandwidth = config.lane_count * config.bit_rate_mbps
-    
+
     for width, height, name in resolutions:
         frame_bits = width * height * 8  # RAW8 format
         overhead_factor = 1.1  # Account for packet headers and timing
         required_bandwidth = frame_bits * overhead_factor
-        
+
         max_fps = (total_bandwidth * 1e6) / required_bandwidth
-        
+
         print(f"  {name} ({width}x{height}): Max {max_fps:.1f} fps @ RAW8")
 
 def test_error_handling():
     """Test error injection and handling capabilities"""
     print("\n=== ERROR HANDLING TESTS ===")
-    
+
     # Test 1: ECC Error Simulation
     print("\n1. ECC Error Simulation")
     original_ecc = calculate_ecc(0x2A, 1024)
-    
+
     # Simulate single-bit errors
     error_patterns = [0x01, 0x02, 0x04, 0x08, 0x10, 0x20, 0x40, 0x80]
-    
+
     for error in error_patterns:
         corrupted_ecc = original_ecc ^ error
         print(f"  Original: 0x{original_ecc:02x}, Corrupted: 0x{corrupted_ecc:02x}, "
               f"Error pattern: 0x{error:02x}")
-    
+
     # Test 2: Configuration Validation
     print("\n2. Configuration Validation")
     invalid_configs = [
@@ -202,7 +202,7 @@ def test_error_handling():
         {"trio_count": 0, "error": "Invalid trio count"},
         {"trio_count": 5, "error": "Too many trios"}
     ]
-    
+
     for config_params in invalid_configs:
         error_desc = config_params.pop("error")
         try:
@@ -217,28 +217,28 @@ def test_error_handling():
 def test_virtual_channels():
     """Test virtual channel functionality"""
     print("\n=== VIRTUAL CHANNEL TESTS ===")
-    
+
     # Test 1: Standard Virtual Channels
     print("\n1. Standard Virtual Channels (CSI-2 v1.x)")
     standard_vcs = [VirtualChannel.VC0, VirtualChannel.VC1, VirtualChannel.VC2, VirtualChannel.VC3]
-    
+
     for vc in standard_vcs:
         packet = Csi2ShortPacket.frame_start(vc, 1)
         print(f"  VC{vc.value}: DI=0x{packet.header.data_id:02x}, "
               f"VC_extracted={packet.virtual_channel}")
-    
-    # Test 2: Extended Virtual Channels  
+
+    # Test 2: Extended Virtual Channels
     print("\n2. Extended Virtual Channels (CSI-2 v2.0+)")
     extended_vcs = [VirtualChannel.VC4, VirtualChannel.VC8, VirtualChannel.VC12, VirtualChannel.VC15]
-    
+
     for vc in extended_vcs:
         # Extended VCs require special handling in CSI-2 v2.0+
         print(f"  VC{vc.value}: Extended VC support required")
-    
+
     # Test 3: Multi-VC Packet Streams
     print("\n3. Multi-VC Packet Simulation")
     packet_sequences = []
-    
+
     for vc in range(4):
         # Create a typical frame sequence for each VC
         sequence = [
@@ -249,21 +249,21 @@ def test_virtual_channels():
             Csi2ShortPacket.frame_end(vc, vc + 1)
         ]
         packet_sequences.append(sequence)
-        
+
         total_bytes = sum(p.get_packet_length() for p in sequence)
         print(f"  VC{vc} sequence: {len(sequence)} packets, {total_bytes} bytes total")
 
 def test_advanced_features():
     """Test advanced CSI-2 features"""
     print("\n=== ADVANCED FEATURES ===")
-    
+
     # Test 1: Scrambling Configuration
     print("\n1. Scrambling Support (CSI-2 v2.0+)")
     scrambling_configs = [
         (False, "Disabled"),
         (True, "Enabled")
     ]
-    
+
     for enabled, status in scrambling_configs:
         config = Csi2Config(
             phy_type=PhyType.DPHY,
@@ -272,14 +272,14 @@ def test_advanced_features():
             scrambling_enabled=enabled
         )
         print(f"  Scrambling {status}: {config.scrambling_enabled}")
-    
+
     # Test 2: Continuous vs Non-Continuous Clock
     print("\n2. Clock Mode Configuration")
     clock_modes = [
         (True, "Continuous", "Clock always active"),
         (False, "Non-continuous", "Clock gated during LP mode")
     ]
-    
+
     for mode, name, description in clock_modes:
         config = Csi2Config(
             phy_type=PhyType.DPHY,
@@ -288,33 +288,33 @@ def test_advanced_features():
             continuous_clock=mode
         )
         print(f"  {name}: {description}")
-    
+
     # Test 3: Lane Distribution
     print("\n3. Lane Distribution Features")
     multi_lane_configs = [2, 4, 8]
-    
+
     for lanes in multi_lane_configs:
         if lanes > 4:  # Skip configurations not commonly tested
             continue
-            
+
         config = Csi2Config(
             phy_type=PhyType.DPHY,
             lane_count=lanes,
             bit_rate_mbps=2000,
             lane_distribution_enabled=True
         )
-        
+
         # Calculate bytes per lane for a sample packet
         sample_payload = generate_test_pattern(64, 32, "ramp")  # 2KB
         bytes_per_lane = len(sample_payload) / lanes
-        
+
         print(f"  {lanes}-lane distribution: {bytes_per_lane:.1f} bytes per lane "
               f"for {len(sample_payload)} byte payload")
 
 def test_realistic_scenarios():
     """Test realistic camera interface scenarios"""
     print("\n=== REALISTIC SCENARIOS ===")
-    
+
     # Test 1: Camera Initialization Sequence
     print("\n1. Camera Initialization Sequence")
     init_sequence = [
@@ -324,10 +324,10 @@ def test_realistic_scenarios():
         "Start video stream transmission",
         "Monitor for Frame End packets"
     ]
-    
+
     for i, step in enumerate(init_sequence, 1):
         print(f"  Step {i}: {step}")
-    
+
     # Test 2: Image Sensor Configurations
     print("\n2. Common Image Sensor Configurations")
     sensor_configs = [
@@ -339,7 +339,7 @@ def test_realistic_scenarios():
             "lanes": 2
         },
         {
-            "name": "5MP Sensor", 
+            "name": "5MP Sensor",
             "resolution": (2560, 1920),
             "format": DataType.RAW12,
             "fps": 15,
@@ -353,7 +353,7 @@ def test_realistic_scenarios():
             "lanes": 4
         }
     ]
-    
+
     for sensor in sensor_configs:
         width, height = sensor["resolution"]
         pixel_bits = {
@@ -362,14 +362,14 @@ def test_realistic_scenarios():
             DataType.RAW12: 12,
             DataType.RAW16: 16
         }.get(sensor["format"], 8)
-        
+
         frame_bits = width * height * pixel_bits
         required_bw = frame_bits * sensor["fps"] * 1.1  # 10% overhead
         bw_per_lane = required_bw / sensor["lanes"] / 1e6
-        
+
         print(f"  {sensor['name']}: {width}x{height} @ {sensor['fps']}fps, "
               f"{sensor['lanes']} lanes, {bw_per_lane:.0f} Mbps/lane")
-    
+
     # Test 3: Multi-Camera System
     print("\n3. Multi-Camera System Configuration")
     camera_system = [
@@ -378,28 +378,28 @@ def test_realistic_scenarios():
         {"id": 2, "vc": 2, "res": (160, 120), "purpose": "Thumbnail camera"},
         {"id": 3, "vc": 3, "res": (80, 60), "purpose": "Status indicator"}
     ]
-    
+
     total_bandwidth = 0
     for camera in camera_system:
         width, height = camera["res"]
         frame_rate = 30
         bits_per_pixel = 8
-        
+
         camera_bw = width * height * bits_per_pixel * frame_rate
         total_bandwidth += camera_bw
-        
+
         print(f"  Camera {camera['id']} (VC{camera['vc']}): {width}x{height}, "
               f"{camera['purpose']}, {camera_bw/1e6:.1f} Mbps")
-    
+
     print(f"  Total system bandwidth: {total_bandwidth/1e6:.1f} Mbps")
 
 def main():
     """Run comprehensive CSI-2 extension tests"""
     print("MIPI CSI-2 Extension Comprehensive Test Suite")
     print("=" * 50)
-    
+
     start_time = time.time()
-    
+
     try:
         test_protocol_compliance()
         test_data_formats()
@@ -408,22 +408,22 @@ def main():
         test_virtual_channels()
         test_advanced_features()
         test_realistic_scenarios()
-        
+
         end_time = time.time()
         duration = end_time - start_time
-        
+
         print(f"\n{'='*50}")
         print("ALL TESTS COMPLETED SUCCESSFULLY")
         print(f"Test duration: {duration:.2f} seconds")
         print(f"CSI-2 extension validation: PASSED")
         print("Ready for production use and HDL simulation")
-        
+
     except Exception as e:
         print(f"\nTest failed with error: {e}")
         import traceback
         traceback.print_exc()
         return 1
-    
+
     return 0
 
 if __name__ == "__main__":
